add coverage reporting (#3)
* add coverage reporting

* fix coverage path to be the installed package dir

* improve help test

* capture coverage report in ci

* update readme
