+++ .coveragerc
@@ -1,0 +1,11 @@
+[paths]
+source =
+  src/
+
+[report]
+precision = 2
+
+[run]
+branch = True
+source =
+  src/
