+++ pytest.ini
@@ -1,5 +1,6 @@
 [pytest]
 addopts =
-    -vv
+    --cov-report=html
     --showlocals
+    -vv
 testpaths = tests
