+++ requirements.test.txt
@@ -5,3 +5,4 @@ flake8-import-order~=0.18
 pdbpp~=0.9
 pep8-naming~=0.7
 pytest~=3.8
+pytest-cov~=2.6
