+++ tox.ini
@@ -3,7 +3,7 @@ envlist = py35,py36,py37,flake8,linters
 
 [testenv]
 deps = .[test]
-commands = {posargs:pytest}
+commands = pytest --cov={envsitepackagesdir}/ttc_api_scraper {posargs}
 
 # Linters
 [testenv:flake8]
